implemented entire Deck class for reading files and dealing cards. Changed animal lifespan field to float type.

+++ Animal.java
@@ -4,10 +4,10 @@ public class Animal {
     private float size;
     private float weight;
     private float speed;
-    private int lifeSpan;
+    private float lifeSpan;
 
 
-    public Animal(String name, float size, float weight, float speed, int lifeSpan) {
+    public Animal(String name, float size, float weight, float speed, float lifeSpan) {
         this.name = name;
         this.size = size;
         this.weight = weight;
@@ -36,7 +36,7 @@ public class Animal {
     }
 
 
-    public int getLifeSpan() {
+    public float getLifeSpan() {
         return lifeSpan;
     }
 
@@ -44,7 +44,7 @@ public class Animal {
     public String toString() {
         String line = "-----------------------------------------------";
         return String.format("%.25s \n| %-21s |\n%.25s \n| Size: %12.2f cm |"
-        + "\n| Weight: %10.2f kg |\n| Top speed: %6.2f kph |\n| Life span: %4d years |"
+        + "\n| Weight: %10.2f kg |\n| Top speed: %6.2f kph |\n| Life span: %4.0f years |"
         + " \n%.25s", line, name.toUpperCase(), line, size, weight, speed, lifeSpan, line);
     }
 }
